config/config.py
import json
import os
import threading
import tkinter as tk
from tkinter import ttk,filedialog
from PIL import Image, ImageTk
import albumentations as A
from albumentations.pytorch import ToTensorV2
from pytorch_lightning.callbacks import ModelCheckpoint
from torch.utils.data import DataLoader
from matplotlib import pyplot as plt
import torch
import wandb
import numpy as np
import random
import gc
from tqdm import tqdm
import scipy.stats as st
from pytorch_lightning.loggers import WandbLogger