requirement.txt
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
appdirs==1.4.4
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
efficientnet_pytorch==0.7.1
filelock==3.9.0
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.41
huggingface-hub==0.17.3
idna==3.4
imageio==2.33.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
munch==4.0.0
networkx==3.0
numpy==1.24.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
Pillow==9.3.0
pretrainedmodels==0.7.4
protobuf==4.25.2
psutil==5.9.8
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.3
PyYAML==6.0.1
qudida==0.0.4
requests==2.28.1
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
segmentation-models-pytorch==0.3.3
sentry-sdk==1.39.2
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.9.2
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchmetrics==1.3.0.post0
torchvision==0.16.2+cu121
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.16.2
warmup_scheduler==0.3
yarl==1.9.4
